Completed reorganisation of source code hierarchy. All source code now falls under the src directory Compilation can be done from the root directory, or individual application dirs
Makefile system completely rewritten to make management much easier!


git-svn-id: https://www.radio.warwick.ac.uk/repos/dps/trunk@56 156110d5-7a05-0410-a37f-c1950ce4f093

+++ src/audio/InputRaw.cpp
@@ -30,7 +30,7 @@ Audio::InputRaw::~InputRaw() {
 
 }
 
-void Audio::InputRaw::getAudio(short *audioData, int samples) {
+void Audio::InputRaw::getAudio(short *audioData, unsigned long samples) {
     char *ptr = (char*)audioData;
     short count = 0;
     long x;
@@ -138,7 +138,7 @@ void Audio::InputRaw::removeCounter(Audio::Counter *counter) {
 }
 
 void Audio::InputRaw::updateCounters(long sample) {
-	for (int i = 0; i < countersList.size(); i++) {
+	for (unsigned int i = 0; i < countersList.size(); i++) {
 		//countersList.at(i)->update(sample);
 	}
 }
